added file input and output from/to yaml files

+++ src/main/java/cz/cvut/fel/pjv/model/EnemyShip.java
@@ -7,8 +7,8 @@ import javafx.scene.image.Image;
  * Enemy Ship object, which can shoot projectile
  */
 public class EnemyShip extends Ship {
-	public EnemyShip(SpaceExplorationEngine spaceExplorationEngine, double positionX, double positionY, double velocityX, double velocityY, String spriteBound, double life, double damage, Projectile projectile, Image... spriteImage) {
-		super(spaceExplorationEngine,positionX, positionY, velocityX, velocityY, spriteBound, life, damage, projectile, spriteImage);
+	public EnemyShip(SpaceExplorationEngine spaceExplorationEngine, double positionX, double positionY, double velocityX, double velocityY, String spriteBound, int life, double damage, Projectile projectile, Image... spriteImage) {
+		super(spaceExplorationEngine, positionX, positionY, velocityX, velocityY, spriteBound, life, damage, projectile, spriteImage);
 	}
 
 	@Override
